cli-python/requirements.txt
python-dotenv==1.2.1
requests==2.32.5
PyPDF2==3.0.1
PyYAML>=6.0
